+++ src/main/java/com/taskManager/usuario/Usuario.java
@@ -16,7 +16,7 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUsuario;
+    private Long idUsuario;
 
     @Column(unique = true)
     private String username;
